+++ commitlint.config.mjs
@@ -7,6 +7,7 @@
  */
 export default {
   extends: ["@commitlint/config-conventional"],
+  ignores: [(message) => /^Signed-off-by: dependabot\[bot\]/m.test(message)],
   rules: {
     "type-enum": [
       2,